LayoutTests/imported/w3c/web-platform-tests/intersection-observer/target-in-different-window-expected.txt


FAIL IntersectionObserver with target in a different window. assert_equals: Notification after insertion into popup. expected 2 but got 1
